feat: basic Address CRUD

+++ src/main/java/com/lsoria/qrsec/controller/rest/AddressController.java
@@ -1,13 +1,17 @@
 package com.lsoria.qrsec.controller.rest;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 import com.lsoria.qrsec.domain.dto.AddressDTO;
 import com.lsoria.qrsec.domain.dto.mapper.AddressMapper;
 import com.lsoria.qrsec.domain.model.Address;
+import com.lsoria.qrsec.domain.model.Role;
+import com.lsoria.qrsec.domain.model.User;
 import com.lsoria.qrsec.service.AddressService;
+import com.lsoria.qrsec.service.UserService;
 import com.lsoria.qrsec.service.exception.ConflictException;
 import com.lsoria.qrsec.service.exception.NotFoundException;
 
@@ -32,6 +36,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -45,10 +50,23 @@ public class AddressController {
     AddressService addressService;
 
     @Autowired
+    UserService userService;
+
+    @Autowired
     AddressMapper addressMapper;
 
     @Operation(summary = "Get all Addresses (privileged)", description = "Get all Addresses from the neighbourhood")
-    @GetMapping("${api.path.addresses}")
+    @GetMapping("${api.path.admin.addresses}")
+    @Parameter(
+            name = "X-Email",
+            description = "Email of the Admin that wants to see the Addresses",
+            in = ParameterIn.HEADER,
+            required = true,
+            schema = @Schema(
+                    type = "string",
+                    example = "exa@mple.com"
+            )
+    )
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
@@ -66,22 +84,71 @@ public class AddressController {
                     content = @Content()
             ),
             @ApiResponse(
+                    responseCode = "401",
+                    description = "User is not Authorized (not a privileged User)",
+                    content = @Content()
+            ),
+            @ApiResponse(
+                    responseCode = "404",
+                    description = "User not found on the database",
+                    content = @Content()
+            ),
+            @ApiResponse(
                     responseCode = "500",
                     description = "Some error prevented the Addresses from being retrieved",
                     content = @Content()
             )
     })
-    // TODO: Add @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<List<AddressDTO>> getAddresses() {
+    public ResponseEntity<List<AddressDTO>> getAddresses(
+            @RequestHeader(value = "X-Email") @NotNull String email
+    ) {
 
-        List<Address> addresses = addressService.findAll();
+        try {
 
-        return ResponseEntity.ok(addresses.stream().map(addressMapper::addressToAddressDTO).collect(Collectors.toList()));
+            // TODO: Replace with @PreAuthorize("hasAuthority('ADMIN')")
+            if (!userService.userIsAuthorized(email, new Role(Role.ADMIN))) {
+
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
+            }
+
+            List<Address> addresses = addressService.findAll();
+            if (addresses.isEmpty()) {
+
+                return ResponseEntity.noContent().build();
+
+            }
+
+            return ResponseEntity.ok(addresses.stream().map(addressMapper::addressToAddressDTO).collect(Collectors.toList()));
+
+        } catch (NotFoundException exception) {
+
+            log.error("Message: {}.", exception.getMessage());
+
+            return ResponseEntity.notFound().build();
+
+        } catch (Exception exception) {
+
+            log.error("Couldn't find all the Addresses.\nMessage: {}.\nStackTrace:\n{}", exception.getMessage(), exception.getStackTrace());
+
+        }
+
+        return ResponseEntity.internalServerError().build();
 
     }
 
     @Operation(summary = "Get an Address (privileged)", description = "Get an specific Address")
     @GetMapping("${api.path.addresses}/{id}")
+    @Parameter(
+            name = "X-Email",
+            description = "Email of the User that wants to see an Address",
+            in = ParameterIn.HEADER,
+            required = true,
+            schema = @Schema(
+                    type = "string",
+                    example = "exa@mple.com"
+            )
+    )
     @Parameter(
             name = "id",
             description = "Address uuid",
@@ -103,6 +170,11 @@ public class AddressController {
                     )
             ),
             @ApiResponse(
+                    responseCode = "401",
+                    description = "User is not Authorized",
+                    content = @Content()
+            ),
+            @ApiResponse(
                     responseCode = "404",
                     description = "Address not found",
                     content = @Content()
@@ -113,25 +185,64 @@ public class AddressController {
                     content = @Content()
             )
     })
-    // TODO: Add @PreAuthorize("hasAuthority('ADMIN')")
     public ResponseEntity<AddressDTO> getAddress(
+            @RequestHeader(value = "X-Email") @NotNull String email,
             @PathVariable @NotNull String id
     ) {
 
-        Optional<Address> address = addressService.findOne(id);
+        try {
 
-        if (address.isEmpty()) {
+            // TODO: Replace with @PreAuthorize("hasAuthority('OWNER') or hasAuthority('ADMIN') or hasAuthority('GUARD')")
+            Optional<User> currentUser = userService.findByUsername(email);
+            if (currentUser.isEmpty()) {
+
+                return ResponseEntity.notFound().build();
+
+            }
+            Optional<Address> address = addressService.findOne(id);
+            if (address.isEmpty()) {
+
+                return ResponseEntity.notFound().build();
+
+            }
+            if ((currentUser.get().getAuthorities() == null || currentUser.get().getAuthorities().isEmpty()) || (userService.userIsAuthorized(email, new Role(Role.OWNER)) && !Objects.equals(address.get(), currentUser.get().getAddress()))) {
+
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
+            }
+
+            AddressDTO addressDTO = addressMapper.addressToAddressDTO(address.get());
+
+            return ResponseEntity.ok(addressDTO);
+
+        } catch (NotFoundException exception) {
+
+            log.error("Message: {}.", exception.getMessage());
+
             return ResponseEntity.notFound().build();
+
+        } catch (Exception exception) {
+
+            log.error("Couldn't find the Address {}.\nMessage: {}.\nStackTrace:\n{}", id, exception.getMessage(), exception.getStackTrace());
+
         }
 
-        AddressDTO addressDTO = addressMapper.addressToAddressDTO(address.get());
-
-        return ResponseEntity.ok(addressDTO);
+        return ResponseEntity.internalServerError().build();
 
     }
 
     @Operation(summary = "Create an Address", description = "Save an Address for later use on a User")
     @PostMapping("${api.path.addresses}")
+    @Parameter(
+            name = "X-Email",
+            description = "Email of the User that wants to create the Address",
+            in = ParameterIn.HEADER,
+            required = true,
+            schema = @Schema(
+                    type = "string",
+                    example = "exa@mple.com"
+            )
+    )
     @io.swagger.v3.oas.annotations.parameters.RequestBody(
             description = "New Address",
             required = true,
@@ -155,19 +266,37 @@ public class AddressController {
                     content = @Content()
             ),
             @ApiResponse(
+                    responseCode = "401",
+                    description = "User is not Authorized",
+                    content = @Content()
+            ),
+            @ApiResponse(
+                    responseCode = "404",
+                    description = "User not found",
+                    content = @Content()
+            ),
+            @ApiResponse(
                     responseCode = "500",
                     description = "Some error prevented the Address from being created",
                     content = @Content()
             )
     })
     public ResponseEntity<AddressDTO> createAddress(
+            @RequestHeader(value = "X-Email") @NotNull String email,
             @RequestBody @NotNull AddressDTO addressDTO
     ) {
 
-        Address addressToCreate = addressMapper.addressDTOToAddress(addressDTO);
-        addressToCreate.setId(null);
-
         try {
+
+            // TODO: Replace with @PreAuthorize("hasAuthority('OWNER')")
+            if (!userService.userIsAuthorized(email, new Role(Role.OWNER))) {
+
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
+            }
+
+            Address addressToCreate = addressMapper.addressDTOToAddress(addressDTO);
+            addressToCreate.setId(null);
 
             Address createdAddress = addressService.save(addressToCreate);
 
@@ -175,13 +304,13 @@ public class AddressController {
 
         } catch (ConflictException conflictException) {
 
-            log.error("Couldn't create ADDRESS.\nMessage: {}.", conflictException.getMessage());
+            log.error("Message: {}.", conflictException.getMessage());
 
             return ResponseEntity.status(HttpStatus.CONFLICT).build();
 
         } catch (Exception exception) {
 
-            log.error("Couldn't create ADDRESS with values:\n{}\nMessage: {}.\nStackTrace:\n{}", addressDTO, exception.getMessage(), exception.getStackTrace());
+            log.error("Couldn't create Address with values:\n{}\nMessage: {}.\nStackTrace:\n{}", addressDTO, exception.getMessage(), exception.getStackTrace());
 
         }
 
@@ -191,6 +320,16 @@ public class AddressController {
 
     @Operation(summary = "Update an Address", description = "Update address's information")
     @PutMapping("${api.path.addresses}/{id}")
+    @Parameter(
+            name = "X-Email",
+            description = "Email of the User that wants to update the Address",
+            in = ParameterIn.HEADER,
+            required = true,
+            schema = @Schema(
+                    type = "string",
+                    example = "exa@mple.com"
+            )
+    )
     @Parameter(
             name = "id",
             description = "Address uuid",
@@ -220,6 +359,11 @@ public class AddressController {
                     )
             ),
             @ApiResponse(
+                    responseCode = "401",
+                    description = "User is not Authorized",
+                    content = @Content()
+            ),
+            @ApiResponse(
                     responseCode = "404",
                     description = "Address not found",
                     content = @Content()
@@ -231,28 +375,47 @@ public class AddressController {
             )
     })
     public ResponseEntity<AddressDTO> updateAddress(
+            @RequestHeader(value = "X-Email") @NotNull String email,
             @PathVariable @NotNull String id,
             @RequestBody @NotNull AddressDTO addressDTO
     ) {
 
-        Optional<Address> foundAddress = addressService.findOne(id);
-
-        if (foundAddress.isEmpty()) {
-            return ResponseEntity.notFound().build();
-        }
-
-        Address addressToUpdate = foundAddress.get();
-        Address addressNewValues = addressMapper.addressDTOToAddress(addressDTO);
-
         try {
+
+            Optional<User> currentUser = userService.findByUsername(email);
+            if (currentUser.isEmpty()) {
+
+                return ResponseEntity.notFound().build();
+
+            }
+            Optional<Address> foundAddress = addressService.findOne(id);
+            if (foundAddress.isEmpty()) {
+
+                return ResponseEntity.notFound().build();
+
+            }
+            if ((!userService.userIsAuthorized(email, new Role(Role.ADMIN))) && (userService.userIsAuthorized(email, new Role(Role.OWNER)) && !Objects.equals(foundAddress.get(), currentUser.get().getAddress()))) {
+
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
+            }
+
+            Address addressToUpdate = foundAddress.get();
+            Address addressNewValues = addressMapper.addressDTOToAddress(addressDTO);
 
             Address updatedAddress = addressService.update(addressToUpdate, addressNewValues);
 
             return ResponseEntity.ok(addressMapper.addressToAddressDTO(updatedAddress));
 
+        } catch (NotFoundException exception) {
+
+            log.error("Message: {}.", exception.getMessage());
+
+            return ResponseEntity.notFound().build();
+
         } catch (Exception exception) {
 
-            log.error("Couldn't update ADDRESS with values:\n{}\nMessage: {}.\nStackTrace:\n{}", addressDTO, exception.getMessage(), exception.getStackTrace());
+            log.error("Couldn't update Address with values:\n{}\nMessage: {}.\nStackTrace:\n{}", addressDTO, exception.getMessage(), exception.getStackTrace());
 
         }
 
@@ -262,6 +425,16 @@ public class AddressController {
 
     @Operation(summary = "Delete an Address", description = "The Address will no longer be available in the database")
     @DeleteMapping("${api.path.addresses}/{id}")
+    @Parameter(
+            name = "X-Email",
+            description = "Email of the Owner that wants to delete the Invite",
+            in = ParameterIn.HEADER,
+            required = true,
+            schema = @Schema(
+                    type = "string",
+                    example = "exa@mple.com"
+            )
+    )
     @Parameter(
             name = "id",
             description = "Address uuid",
@@ -280,6 +453,11 @@ public class AddressController {
                     content = @Content()
             ),
             @ApiResponse(
+                    responseCode = "401",
+                    description = "User is not Authorized",
+                    content = @Content()
+            ),
+            @ApiResponse(
                     responseCode = "404",
                     description = "Address not found",
                     content = @Content()
@@ -291,10 +469,18 @@ public class AddressController {
             )
     })
     public ResponseEntity<AddressDTO> deleteAddress(
+            @RequestHeader(value = "X-Email") @NotNull String email,
             @PathVariable @NotNull String id
     ) {
 
         try {
+
+            // TODO: Replace with @PreAuthorize("hasAuthority('OWNER')")
+            if (!userService.userIsAuthorized(email, new Role(Role.OWNER))) {
+
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
+            }
 
             addressService.delete(id);
 
@@ -302,11 +488,13 @@ public class AddressController {
 
         } catch (NotFoundException exception) {
 
+            log.error("Message: {}.", exception.getMessage());
+
             return ResponseEntity.notFound().build();
 
         } catch (Exception exception) {
 
-            log.error("Couldn't delete ADDRESS {}.\nMessage: {}.\nStackTrace:\n{}", id, exception.getMessage(), exception.getStackTrace());
+            log.error("Couldn't delete the Address {}.\nMessage: {}.\nStackTrace:\n{}", id, exception.getMessage(), exception.getStackTrace());
 
         }
